requirements.txt
Django 
djangorestframework 
django-rest-auth
django-heroku
python-dateutil
psycopg2-binary
gunicorn
cryptography
django-extensions
djangorestframework-jwt
python-jose~=3.0
python-dotenv~=0.9